docs(scalars): add stories for the textarea
